Add availableChunks support for static exports.
